requirements.txt
streamlit>=1.27.0
pymupdf>=1.22.0
pandas>=2.0.0
python-docx>=0.8.11
pytesseract>=0.3.10
opencv-python>=4.7.0
Pillow>=10.0.0
ollama>=0.1.0 